power

+++ 0x08-recursion/4-pow_recursion.c
@@ -10,6 +10,6 @@
 int _pow_recursion(int x, int y)
 {
 if (y < 0) return (-1);
-double i = pow(x,y);
+int i = pow(x,y);
 return (i);
 }
